requirements.txt
cryptography
ansible
ansible-lint
paramiko
jsondiff
hcloud
hetzner
jmespath
passlib
ara
kubernetes
jsonpatch
PyYAML
netaddr
azure-keyvault
azure-identity
